requirements.txt
Flask
Flask-MySQLdb
Flask-Login
Werkzeug
PyMySQL
python-dotenv